Introduce rseq-abi.h
Introduce our own header for the rseq kernel ABI, so we do not depend on
Linux kernel headers.

Use this header's rseq_cs.arch.ptr field as accessor for the rseq_cs ptr
on 32-bit.

Remove the extra/ directory because the copy of the Linux kernel headers
is not needed anymore.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I0170f2bcaa7defd3a0d3f9a38115bc7d321ddde2