param test: membarrier: validate total number of increments
While reviewing the riscv implementation for this test, I noticed that
the final store is only incrementing the content of its register,
without ever storing it to memory.

This passes testing because the increment is effectively a no-op, and
the test never validates that any increment happen in the first place.

Introduce a validation of the increment total to eliminate those
false-positives.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ie9ee1c7902509da75dd85babd07bc180ad0b9ae9