Only set 'rseq_size' on first thread registration
To mimic the libc behavior, don't set the feature size in 'rseq_size'
until at least one thread has successfully registered. This allows using
'rseq_size' in fast-paths to test for both registration status and
available features. The caveat is that on libc either all threads are
registered or none are, while with bare librseq it is the responsability
of the user to register all threads using rseq.

Change-Id: Idc491153de55f9ac38bd1fcad425aee9549fc910
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>