Fix: tests: Do not expect allowed cpus to be invariant
When running in containers such as Incus, the allowed cpus can change
under the hood. Adapt the basic_percpu_ops_test and param_test tests to
make no assumption about allowed cpus being invariant.

mm_cid_shrink_test only uses set_affinity as a hint. If the affinity is
not preserved, the test is only less likely to generate the scenario
tested for, but it won't trigger a failure.

basic_percpu_benchmark set_affinity only uses affinity as a hint.

basic_test really requires affinity to be unchanged by external
processes. This is an issue with Incus. This is not handled by this
fix.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ia72314c6f46a1b1de98439ac3e2e3f75cecd1e06