Introduce librseq-reg-helper.so
Introduce librseq-reg-helper.so to provide rseq in environment with libc
that does not have rseq support.

librseq-reg-helper.so contains an Initial-Exec TLS variable, so it is
not suitable for dlopen(). It needs to be linked against explicitly by
the program.

rseq_register_current_thread() and rseq_unregister_current_thread() are
now moved to librseq-reg-helper.so.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I157914a59c496ba87288ef094559280d90fea2c8