Remove support for 'standalone' rseq registration
Depend on libc to handle the allocation and registration of the rseq
area for the whole process.

Librseq can still be built and run with a libc that lacks the rseq
symbols as it doesn't link on them directly but it will not handle the
registration even if the syscall is available.

User visible changes:

  * Removes per-thread registration API, with libc the registration is
    either enabled for all threads or none.

Change-Id: Ib0cb9aa8ccc8c8786ed0fdd2a50cebe23e4fb1b1
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>