Improve syscall errors test
Rework the test to first register with the original 32 byte ABI and then
with the current feature size if it's greater. This ensures we test the
'legacy' registration size on newer kernels.

Also add a test that registers an rseq area with a non-null rseq_cs
pointer to a valid structure. This ensures we don't break compatibility
with older libc that reuses rseq areas without clearing the rseq_cs
pointer.

Change-Id: Iefbf1a6fef77e3ae7bf1218ecff77b74a56dc180
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>