Use RCU for hierarchical byte counters precise sum
The precise sum is only accurate if updaters are quiescent, because they
may be in flight and propagating a carry.

Surround updaters by an RCU read-side critical section, and use the
level 0 bit mask value 0 to divert updaters to the global approximate
counter sum while at least one precise sum is being aggregated.

Precise sum aggregation wait for a grace period to ensure updaters
are not modifying the level 0 and intermediate tree nodes while the
precise sum reads those values.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ie5ffe1030fe4f899002bf05577c6b1c03a1318da